made protocol_login work
  I added a `generic` interface for ringbuffers and specifically tasks
ringbuffers which will be most of its use cases. This made me consider
changing back to C++ for these kind of containers and perhaps some
unique_ptr's as I ended up using heap allocations for the login_info.
  I ended up choosing C because I didn't want any `under the hood`
allocation but it might simplify some code if we use a minimal set of
C++.